KYC completed 08/12/2025

+++ app/src/main/java/com/bos/payment/appName/data/model/travel/bus/busSeatMap/FareMaster.kt
@@ -5,15 +5,15 @@ import com.google.gson.annotations.SerializedName
 
 data class FareMaster (
 
-  @SerializedName("basic_Amount"         ) var basicAmount         : Int?                   = null,
-  @SerializedName("cancellation_Charges" ) var cancellationCharges : Int?                   = null,
+  @SerializedName("basic_Amount"         ) var basicAmount         : Double?                   = null,
+  @SerializedName("cancellation_Charges" ) var cancellationCharges : Double?                   = null,
   @SerializedName("fareDetails"          ) var fareDetails         : ArrayList<FareDetails> = arrayListOf(),
-  @SerializedName("gst"                  ) var gst                 : Int?                   = null,
-  @SerializedName("gross_Commission"     ) var grossCommission     : Int?                   = null,
-  @SerializedName("net_Commission"       ) var netCommission       : Int?                   = null,
+  @SerializedName("gst"                  ) var gst                 : Double?                   = null,
+  @SerializedName("gross_Commission"     ) var grossCommission     : Double?                   = null,
+  @SerializedName("net_Commission"       ) var netCommission       : Double?                   = null,
   @SerializedName("other_Amount"         ) var otherAmount         : Double?                = null,
-  @SerializedName("service_Fee_Amount"   ) var serviceFeeAmount    : Int?                   = null,
+  @SerializedName("service_Fee_Amount"   ) var serviceFeeAmount    : Double?                   = null,
   @SerializedName("total_Amount"         ) var totalAmount         : Double?                = null,
-  @SerializedName("trade_Markup_Amount"  ) var tradeMarkupAmount   : Int?                   = null
+  @SerializedName("trade_Markup_Amount"  ) var tradeMarkupAmount   : Double?                   = null
 
 )
